requirements.txt
asgiref==3.2.10
autopep8==1.5.4
coverage==5.2.1
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
django-mysql==3.9
django-environ==0.8.1
django-cors-headers==3.5.0
pika==1.1.0
pycodestyle==2.6.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1

django-extensions==3.1.5
pydot ==1.4.2
django-rest-swagger==2.2.0
drf-yasg==1.20.0

pytest-django==4.4.0
pyparsing==3.0.7
pytest==7.1.1
factory-boy==3.2.1
Faker==8.9.1
pytest-cov==2.12.1
pytest-factoryboy==2.1.0
pytest-django==4.4.0
